src/requirements.txt
# Versões ultra-estáveis para máxima compatibilidade, evitando erros de compilação
pandas==1.5.3
numpy==1.24.4
openpyxl==3.1.2
selenium==4.15.2
webdriver-manager==4.0.1
psutil==5.9.6
Pillow==10.1.0
